lift planting complete update dispatch out of dialog
